Cleanup controller/middleware distinction.
