+++ app/controllers/influencer.js
